Move runner to its own folder
