+++ crates/karva_core/src/package.rs
@@ -120,7 +120,7 @@ impl<'proj> Package<'proj> {
     pub fn total_test_modules(&self) -> usize {
         let mut total = 0;
         for module in self.modules.values() {
-            if module.module_type == ModuleType::Test {
+            if module.total_test_cases() > 0 {
                 total += 1;
             }
         }
